FL_Healthcare_Canada/requirements.txt
# FL Healthcare Canada - Requirements
# Complete dependencies for federated learning with AI governance

# ============================================================================
# Core Deep Learning
# ============================================================================
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# ============================================================================
# Numerical Computing & Data Processing
# ============================================================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.11.0,<2.0.0

# ============================================================================
# Machine Learning & Evaluation
# ============================================================================
scikit-learn>=1.3.0,<2.0.0

# ============================================================================
# Visualization
# ============================================================================
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# ============================================================================
# Medical Imaging Dataset
# ============================================================================
medmnist>=2.2.0
Pillow>=10.0.0

# ============================================================================
# Document Generation (from evaluation/experiments modules)
# ============================================================================
python-docx>=0.8.11,<1.0.0
openpyxl>=3.1.0,<4.0.0

# ============================================================================
# Configuration & Serialization
# ============================================================================
pyyaml>=6.0,<7.0

# ============================================================================
# Progress Bars & Utilities
# ============================================================================
tqdm>=4.66.0,<5.0.0

# ============================================================================
# Optional: Jupyter Support (for interactive analysis)
# ============================================================================
# jupyter>=1.0.0
# ipykernel>=6.25.0
# notebook>=7.0.0

# ============================================================================
# Optional: Advanced Privacy Libraries
# ============================================================================
# opacus>=1.4.0  # Facebook's DP library for PyTorch

# ============================================================================
# Optional: Distributed Training
# ============================================================================
# ray>=2.7.0  # For distributed FL
# redis>=5.0.0  # For coordination

# ============================================================================
# Development Dependencies (uncomment for development)
# ============================================================================
# pytest>=7.4.0  # Testing
# pytest-cov>=4.1.0  # Coverage
# black>=23.7.0  # Code formatting
# flake8>=6.1.0  # Linting
# mypy>=1.5.0  # Type checking
# sphinx>=7.2.0  # Documentation
